Move action rendering to uiutils
